crud apoiador 2

+++ hbsfinan/src/main/java/com/hbs/hbsfinan/model/Apoiador.java
@@ -11,7 +11,7 @@ import java.util.Date;
 public class Apoiador {
 
     @Id
-    private int Id;
+    private Long Id;
 
     private String nome;
     private String email;
@@ -23,7 +23,7 @@ public class Apoiador {
 
     public Apoiador(){}
 
-    public Apoiador(String nome, String email, String fone, String endereco, String sexo, String cpf, Date dataNasc, int Id)
+    public Apoiador(String nome, String email, String fone, String endereco, String sexo, String cpf, Date dataNasc, Long Id)
     {
         this.cpf=cpf;
         this.dataNasc=dataNasc;
@@ -59,7 +59,7 @@ public class Apoiador {
         this.fone = fone;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         Id = id;
     }
 
@@ -76,7 +76,7 @@ public class Apoiador {
         return dataNasc;
     }
 
-    public int getId() {
+    public Long getId() {
         return Id;
     }
